+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Make iconfont task function return a `Promise`
+- feat: Make watch task return a node `FSWatcher` object
 
 ## 0.4.0